chore: remove management apis from hero core
