--- core/connections/ConnectionToHeroApiCore.ts
@@ -1,4 +1,0 @@
-import { ConnectionToCore } from '@ulixee/net';
-import { IHeroCoreApiHandlers } from '../apis';
-
-export default class ConnectionToHeroApiCore extends ConnectionToCore<IHeroCoreApiHandlers, {}> {}
