+++ core/interfaces/ICommandWithResult.ts
@@ -2,8 +2,8 @@ import ICommandMeta from '@ulixee/hero-interfaces/ICommandMeta';
 
 export default interface ICommandWithResult extends ICommandMeta {
   label: string;
-  startTime: number;
-  runtimeMs: number;
+  startTime?: number;
+  runtimeMs?: number;
   isError: boolean;
   frameIdPath?: string;
   resultNodeIds?: number[];
